requirements.txt
# Core framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.0
asyncpg==0.29.0
psycopg2-binary==2.9.9

# Data processing
pandas==2.1.3
numpy==1.25.2
pypdf2==3.0.1
python-multipart==0.0.6
beautifulsoup4==4.12.2
markdownify==0.11.6

# AI/ML Libraries
openai==1.3.7
azure-openai==1.3.0
boto3==1.34.0
transformers==4.36.0
sentence-transformers==2.2.2
langchain==0.0.350
langchain-openai==0.0.2
chromadb==0.4.18
faiss-cpu==1.7.4

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Async & Background Tasks
celery==5.3.4
redis==5.0.1
httpx==0.25.2

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
pre-commit==3.6.0

# Deployment
docker==6.1.3
kubernetes==28.1.0